requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.1
djangorestframework==3.6.2
djangorestframework-xml==1.3.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
gunicorn==19.6.0
psycopg2==2.6.2